RMI assignment client-end
1, re-write Post class. Could work with GUI.
2, Customer class becomes Interface. Add CustomerImpl class. Delete User
class.
3, Change Item.price from BigDecimal to double. Double should be enough
for using.
4, Move ItemTuple class from utils to post package. utils package could
be diacarded some day.

+++ src/store/Item.java
@@ -1,4 +1,6 @@
 package store;
+
+
 /*
  * File:         [Item]
  * Created:      [2015/2/3]
@@ -7,17 +9,18 @@ package store;
  * 
  * Changelog:
  * 2015/2/3 Creation of 'Item'
+ * 2015/2/16 price type from BigDecimal to double
  */
-import java.math.BigDecimal;
+
 
 public class Item {
 
 	private String upc;
 	private String description;
 	private String gif;
-	private BigDecimal price;
+	private double price;
 
-	public Item(String upc, String description, String gif, BigDecimal price) {
+	public Item(String upc, String description, String gif, double price) {
 		this.setUpc(upc);
 		this.setDescription(description);
 		this.setGif(gif);
@@ -49,11 +52,11 @@ public class Item {
 
 	}
 
-	public BigDecimal getPrice() {
+	public double getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price2) {
+	public void setPrice(double price2) {
 		this.price = price2;
 	}
 }
